fix extractors

+++ part4/bloglist/controllers/blogs.js
@@ -1,9 +1,7 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 const config = require('../utils/config')
-const { tokenExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -23,7 +21,8 @@ blogRouter.get('/:id', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
-  const token = tokenExtractor(request)
+  const token = request.token
+  const user = request.user
   const decodedToken = jwt.verify(token, config.SECRET)
 
   if (!decodedToken.id) {
@@ -31,8 +30,6 @@ blogRouter.post('/', async (request, response) => {
       .status(401)
       .json({ error: 'token missing or invalid' })
   }
-
-  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -64,8 +61,26 @@ blogRouter.put('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  const user = request.user
+  const decodedToken = jwt.verify(token, config.SECRET)
+
+  if (!decodedToken.id) {
+    return response
+      .status(401)
+      .json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response
+      .status(401)
+      .json({ error: 'only creator can delete note' })
+  }
 })
 
 module.exports = blogRouter
